requirements.txt
homeassistant>=2024.5.5
paho_mqtt>=2.1.0
protobuf>=5.29.1
jsonpath-ng>=1.7.0
pdfminer.six>=20201018
requests>=2.0.0
